Revert "Switch to NewGCMWithRandomNonce for fips (#7474) (#7549)"
This reverts commit 88946e8454fed3348c56f082cd4406b520aac05d.
