+++ internal/pkg/agent/vault/aesgcm/aesgcm.go
@@ -5,8 +5,11 @@
 package aesgcm
 
 import (
+	"crypto/aes"
+	"crypto/cipher"
 	"crypto/rand"
 	"encoding/hex"
+	"syscall"
 )
 
 // AESKeyType indicates the AES key length.
@@ -52,6 +55,29 @@ func NewKeyHexString(kt AESKeyType) (string, error) {
 	return hex.EncodeToString(key), nil
 }
 
+// Encrypt encrypts the data with AES-GCM
+func Encrypt(key, data []byte) ([]byte, error) {
+	block, err := aes.NewCipher(key)
+	if err != nil {
+		return nil, err
+	}
+
+	aesGCM, err := cipher.NewGCM(block)
+	if err != nil {
+		return nil, err
+	}
+
+	nonce := make([]byte, aesGCM.NonceSize())
+	if _, err = rand.Read(nonce); err != nil {
+		return nil, err
+	}
+
+	// The first parameter is nonce in order to get the ciphertext as concatenation of nonce and encrypted data
+	ciphertext := aesGCM.Seal(nonce, nonce, data, nil)
+
+	return ciphertext, nil
+}
+
 // EncryptHex encrypts with hex string key, producing hex encoded result
 func EncryptHex(key string, data []byte) (string, error) {
 	bkey, err := hex.DecodeString(key)
@@ -63,6 +89,27 @@ func EncryptHex(key string, data []byte) (string, error) {
 		return "", err
 	}
 	return hex.EncodeToString(enc), nil
+}
+
+// Decrypt decrypts the data with AES-GCM
+func Decrypt(key, data []byte) ([]byte, error) {
+	block, err := aes.NewCipher(key)
+	if err != nil {
+		return nil, err
+	}
+
+	aesGCM, err := cipher.NewGCM(block)
+	if err != nil {
+		return nil, err
+	}
+
+	nonceSize := aesGCM.NonceSize()
+	if len(data) < nonceSize {
+		return nil, syscall.EINVAL
+	}
+	nonce, ciphertext := data[:nonceSize], data[nonceSize:]
+
+	return aesGCM.Open(nil, nonce, ciphertext, nil)
 }
 
 // DecryptHex decrypts with hex string key and data
