--- internal/pkg/agent/vault/aesgcm/aesgcm_nofips.go
@@ -1,56 +1,0 @@
-// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
-// or more contributor license agreements. Licensed under the Elastic License 2.0;
-// you may not use this file except in compliance with the Elastic License 2.0.
-
-//go:build !requirefips
-
-package aesgcm
-
-import (
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/rand"
-	"syscall"
-)
-
-// Encrypt encrypts the data with AES-GCM
-func Encrypt(key, data []byte) ([]byte, error) {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return nil, err
-	}
-
-	aesGCM, err := cipher.NewGCM(block)
-	if err != nil {
-		return nil, err
-	}
-
-	nonce := make([]byte, aesGCM.NonceSize())
-	if _, err = rand.Read(nonce); err != nil {
-		return nil, err
-	}
-
-	// The first parameter is nonce in order to get the ciphertext as concatenation of nonce and encrypted data
-	ciphertext := aesGCM.Seal(nonce, nonce, data, nil)
-	return ciphertext, nil
-}
-
-// Decrypt decrypts the data with AES-GCM
-func Decrypt(key, data []byte) ([]byte, error) {
-	block, err := aes.NewCipher(key)
-	if err != nil {
-		return nil, err
-	}
-
-	aesGCM, err := cipher.NewGCM(block)
-	if err != nil {
-		return nil, err
-	}
-
-	nonceSize := aesGCM.NonceSize()
-	if len(data) < nonceSize {
-		return nil, syscall.EINVAL
-	}
-	nonce, ciphertext := data[:nonceSize], data[nonceSize:]
-	return aesGCM.Open(nil, nonce, ciphertext, nil)
-}
