+++ internal/pkg/crypto/io.go
@@ -11,10 +11,35 @@ import (
 	"crypto/pbkdf2"
 	"crypto/rand"
 	"crypto/sha512"
+	"encoding/binary"
 	"errors"
 	"fmt"
 	"io"
 )
+
+// Option is the default options used to generate the encrypt and decrypt writer.
+// NOTE: the defined options need to be same for both the Reader and the writer.
+type Option struct {
+	Generator       bytesGen
+	IterationsCount int
+	KeyLength       int
+	SaltLength      int
+	IVLength        int
+
+	// BlockSize must be a factor of aes.BlockSize
+	BlockSize int
+}
+
+// DefaultOptions is the default options to use when creating the writer, changing might decrease
+// the efficacity of the encryption.
+var DefaultOptions = &Option{
+	IterationsCount: 10000,
+	KeyLength:       32,
+	SaltLength:      64,
+	IVLength:        12,
+	Generator:       randomBytes,
+	BlockSize:       bytes.MinRead,
+}
 
 // versionMagicHeader is the format version that will be added at the beginning of the header and
 // can be used to change how the decryption work in future version.
@@ -97,7 +122,7 @@ func (w *Writer) Write(b []byte) (int, error) {
 			return 0, w.err
 		}
 
-		aesgcm, err := getCipherAEAD(block)
+		aesgcm, err := cipher.NewGCM(block)
 		if err != nil {
 			w.err = fmt.Errorf("could not create the GCM to encrypt: %w", err)
 			return 0, w.err
@@ -133,6 +158,33 @@ func (w *Writer) Write(b []byte) (int, error) {
 	}
 
 	return len(b), nil
+}
+
+func (w *Writer) writeBlock(b []byte) error {
+	// randomly generate the salt and the initialization vector, this information will be saved
+	// on disk in the file as part of the header
+	iv, err := w.generator(w.option.IVLength)
+	if err != nil {
+		w.err = fmt.Errorf("fail to generate random IV: %w", err)
+		return w.err
+	}
+
+	//nolint:errcheck // Ignore the error at this point.
+	w.writer.Write(iv)
+
+	encodedBytes := w.gcm.Seal(nil, iv, b, nil)
+
+	l := make([]byte, 4)
+	binary.LittleEndian.PutUint32(l, uint32(len(encodedBytes))) //nolint:gosec // ignoring unsafe type conversion
+	//nolint:errcheck // Ignore the error at this point.
+	w.writer.Write(l)
+
+	_, err = w.writer.Write(encodedBytes)
+	if err != nil {
+		return fmt.Errorf("fail to encode data: %w", err)
+	}
+
+	return nil
 }
 
 // Reader implements the io.Reader interface and allow to decrypt bytes from the Writer. The reader
@@ -210,7 +262,7 @@ func (r *Reader) Read(b []byte) (int, error) {
 			return 0, r.err
 		}
 
-		aesgcm, err := getCipherAEAD(block)
+		aesgcm, err := cipher.NewGCM(block)
 		if err != nil {
 			r.err = fmt.Errorf("could not create the GCM to decrypt the data: %w", err)
 			return 0, r.err
@@ -246,6 +298,43 @@ func (r *Reader) readTo(b []byte) (int, error) {
 	}
 
 	return n, r.err
+}
+
+func (r *Reader) consumeBlock() error {
+	// Retrieve block information:
+	// - Initialization vector
+	// - Length of the block
+	iv, l, err := r.readBlockInfo()
+	if err != nil {
+		return err
+	}
+
+	encodedBytes := make([]byte, l)
+	_, err = io.ReadAtLeast(r.reader, encodedBytes, l)
+	if err != nil {
+		r.err = fmt.Errorf("fail read the block of %d bytes: %w", l, err)
+	}
+
+	decodedBytes, err := r.gcm.Open(nil, iv, encodedBytes, nil)
+	if err != nil {
+		return fmt.Errorf("fail to decode bytes: %w", err)
+	}
+	r.buf = append(r.buf[:], decodedBytes...)
+
+	return nil
+}
+
+func (r *Reader) readBlockInfo() ([]byte, int, error) {
+	buf := make([]byte, r.option.IVLength+4)
+	_, err := io.ReadAtLeast(r.reader, buf, len(buf))
+	if err != nil {
+		return nil, 0, err
+	}
+
+	iv := buf[0:r.option.IVLength]
+	l := binary.LittleEndian.Uint32(buf[r.option.IVLength:])
+
+	return iv, int(l), nil
 }
 
 // Close will propagate the Close call to the wrapped reader.
