+++ internal/pkg/crypto/io_opts_fips.go
@@ -1,0 +1,32 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
+
+//go:build requirefips
+
+package crypto
+
+import (
+	"errors"
+)
+
+// Validate the options for encoding and decoding values.
+func (o *Option) Validate() error {
+	if o.IVLength < 12 {
+		return errors.New("IVLength must be at least 96 bits (12 bytes)")
+	}
+
+	if o.SaltLength < 16 {
+		return errors.New("SaltLength must be at least 128 bits (16 bytes)")
+	}
+
+	if o.IterationsCount < 1000 {
+		return errors.New("IterationsCount must be at least 1000")
+	}
+
+	if o.KeyLength < 14 {
+		return errors.New("KeyLength must be at least 112 bits (14 bytes)")
+	}
+
+	return nil
+}
