+++ internal/pkg/crypto/io_opts_fips_test.go
@@ -22,8 +22,17 @@ func Test_Option_ValidateFIPS(t *testing.T) {
 		name   string
 		option *Option
 	}{{
+		name: "IVLength is low",
+		option: &Option{
+			IVLength:        10,
+			SaltLength:      20,
+			IterationsCount: 10000,
+			KeyLength:       20,
+		},
+	}, {
 		name: "SaltLength is low",
 		option: &Option{
+			IVLength:        20,
 			SaltLength:      10,
 			IterationsCount: 10000,
 			KeyLength:       20,
@@ -31,6 +40,7 @@ func Test_Option_ValidateFIPS(t *testing.T) {
 	}, {
 		name: "IterationsCount is low",
 		option: &Option{
+			IVLength:        20,
 			SaltLength:      20,
 			IterationsCount: 100,
 			KeyLength:       20,
@@ -38,6 +48,7 @@ func Test_Option_ValidateFIPS(t *testing.T) {
 	}, {
 		name: "KeyLength is low",
 		option: &Option{
+			IVLength:        20,
 			SaltLength:      20,
 			IterationsCount: 10000,
 			KeyLength:       10,
